allow fractional font sizes (ticket 242)

+++ deegree-core/deegree-core-rendering-2d/src/main/java/org/deegree/rendering/r2d/se/parser/PostgreSQLReader.java
@@ -398,7 +398,7 @@ public class PostgreSQLReader {
                 if ( bold != null ) {
                     res.bold = bold;
                 }
-                Integer size = (Integer) rs.getObject( "size" );
+                Double size = (Double) rs.getObject( "size" );
                 if ( size != null ) {
                     res.fontSize = size;
                 }
