+++ deegree-core/deegree-core-rendering-2d/src/main/java/org/deegree/rendering/r2d/styling/components/Font.java
@@ -71,7 +71,7 @@ public class Font implements Copyable<Font> {
     /**
      * Default is 10.
      */
-    public int fontSize = 10;
+    public double fontSize = 10;
 
     /**
      * <code>Style</code>
